inotify: Avoid calling vrele() with a namecache mutex held
In cache_vop_inotify(), we call inotify_log() with a namecache hash lock
held.  inotify_log() looks at all watches registered with the vnode to
see if any of them are interested in the event.  In some cases, we have
to detach and free the watch after logging the event.  This means we
must vrele() the watched vnode, and this must not be done while a
non-sleepable lock held.

Previously, I deferred the vrele() to until the inotify softc and vnode
pollinfo locks were dropped.  However, this is not enough since we may
still be holding the aforementioned namecache lock.  Go further and use
a taskqueue thread to release vnode references.  Introduce a set of
detached watches, and queue a threaded task which releases the vnode
reference.

Reported by:	syzbot+c128f121cb22df95559b@syzkaller.appspotmail.com
Reviewed by:	kib
Fixes:		f1f230439fa4 ("vfs: Initial revision of inotify")
Differential Revision:	https://reviews.freebsd.org/D51685